+++ src/app/task-home/task-home.component.ts
@@ -18,11 +18,11 @@ export class TaskHomeComponent implements OnInit {
   }
 
   onClickDelete(id: number) : void {
-    // confirm ('Are you sure you want to delete this task?');
-    // I need to put an if statement here
+    if (confirm ('Are you sure you want to delete this task?')) {
     this.taskDataService
     .deleteTask(id)
     .subscribe(t => this.getTasks());
+    }
   }
 
   constructor(private taskDataService : TaskDataService) {
