+++ src/app/task-list/task-list.component.ts
@@ -19,11 +19,12 @@ export class TaskListComponent implements OnInit {
 
 
   onClickDelete(id: number): void {
-    // confirm ('Are you sure you want to delete this task?');
-    // I need to put an if statement here
-    this.taskDataService
-    .deleteTask(id)
-    .subscribe(t => this.getTasks());
+    if (confirm ('Are you sure you want to delete this task?')) {
+      this.taskDataService
+      .deleteTask(id)
+      .subscribe(t => this.getTasks());
+    }
+
   }
 
   constructor(private taskDataService : TaskDataService
